Implement best practices on Dockerfile
The MAINTAINER instruction has been deprecated
by Docker in favor of LABEL maintainer="...".

All the apt-related commands should be run in a
single RUN instructions to avoid adding more
layers to the image. Running apt-get clean
and removing temporary files in another
layer does not actually make a smaller image
in the end.

Change-Id: If7c689b0881b6b39d8b88c701dda96dc053b9067